Make install button a normal link
Closes #1358
